Refactor ascent list to seperate component to fix map rerender bug
